Move verification helpers to verifiers.h
Suite runners should include verifiers.h instead of including tester.h
