+++ test/tester.h
@@ -37,66 +37,4 @@ void summarize_tests();
 
 void verify(bool success, const char* msg);
 
-// the following functions and templates are helper functions short-hand to verify
-// call the helpers instead of calling verify directly
-
-inline void is_true(bool value, const char* msg)
-{
-	verify(value, msg);
-}
-
-
-inline void is_false(bool value, const char* msg)
-{
-	verify(!value, msg);
-}
-
-
-//TODO replace this template with a macro to reduce template instantiations
-template <typename T>
-constexpr bool is_nonbool_arithmetic()
-{
-	return std::is_integral_v<T> && !std::is_same_v<T, bool>;
-}
-
-
-template <typename T = int>
-inline void is_zero(T value, const char* msg)
-{
-	static_assert(is_nonbool_arithmetic<T>(), "requires non-bool arithmetic type");
-	verify(value == 0, msg);
-}
-
-
-template <typename T = int>
-inline void is_nonzero(T value, const char* msg)
-{
-	static_assert(is_nonbool_arithmetic<T>(), "requires non-bool arithmetic type");
-	verify(value != 0, msg);
-}
-
-
-template <typename T = int>
-inline void is_negative(T value, const char* msg)
-{
-	static_assert(is_nonbool_arithmetic<T>(), "requires non-bool arithmetic type");
-	verify(value < 0, msg);
-}
-
-
-template <typename T = int>
-inline void is_nonnegative(T value, const char* msg)
-{
-	static_assert(is_nonbool_arithmetic<T>(), "requires non-bool arithmetic type");
-	verify(value >= 0, msg);
-}
-
-
-template <typename T = int>
-inline void is_positive(T value, const char* msg)
-{
-	static_assert(is_nonbool_arithmetic<T>(), "requires non-bool arithmetic type");
-	verify(value > 0, msg);
-}
-
 #endif
